Extract a multi-env test engine into a module.
Add a custom sub-interface of JUnit5 `Extension` for communications between
the new test engine and test extensions that actually manage test execution
environments.

The new test engine is meant to be reused for API v2 tests.

However, this engine is fairly self-contained and it makes sense to keep it
in a dedicated module.

These changes also fix the filter that was meant to exclude multi-version
tests from the Jupiter test engine. It was not sufficient to exclude
intermediate class nodes, leaf nodes have to be excluded too.